+++ selenium/graph_and_variables.test.ts
@@ -2,7 +2,7 @@ const { By, until } = require('selenium-webdriver');
 
 import { getDriver, login, GRAFANA_URL } from './helpers';
 
-describe('configure', () => {
+describe('graph and variables', () => {
   jest.setTimeout(30000);
   let driver;
 
@@ -13,7 +13,7 @@ describe('configure', () => {
     await driver.get(`${GRAFANA_URL}/d/U6rjzWDMz/sine-wave-example`);
     await driver.wait(
       until.elementLocated(
-        By.xpath(`//div[contains(@class, 'navbar-page-btn')]//a[text()[contains(., "Sine Wave Example")]]`)
+        By.xpath(`//a[text()[contains(., "Sine Wave Example")]]`)
       ),
       5 * 1000
     );