+++ selenium/helpers.ts
@@ -1,4 +1,4 @@
-const { By, Builder, until } = require('selenium-webdriver');
+const { By, Builder, error} = require('selenium-webdriver');
 const chromeDriver = require('selenium-webdriver/chrome');
 
 export const GRAFANA_URL = process.env.GRAFANA_URL || 'http://grafana:3000';
@@ -10,7 +10,14 @@ export async function login(driver) {
   await driver.findElement(By.css("input[name='user']")).sendKeys('admin');
   await driver.findElement(By.css("input[name='password']")).sendKeys('admin123');
   await driver.findElement(By.css("button[aria-label='Login button']")).click();
-  await driver.wait(until.elementLocated(By.css('.dashboard-container')), 2 * 1000);
+  await driver.wait(async () => {
+    try {
+      await driver.findElement(By.css("button[aria-label='Login button']"))
+    } catch (err) {
+      if (err instanceof error.NoSuchElementError) return true
+    }
+    return false
+  }, 2 * 1000);
 }
 
 export async function getDriver() {