+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [2.0.2]
+
+This release adds testing against Grafana v8.1.0 and fixes an issue with query variables.
+
+### Fixed
+
+- Query variables can now also be used in Grafana v8.X.X
+
 ## [2.0.1]
 
 This release fixes some long standing issues that prevented the right use of the alerting feature